chore: move profiles manifest files to profiles module
