+++ profiles/manifests/category-attributes.yaml
