+++ profiles/manifests/copy-scope.yaml
