+++ profiles/manifests/full-payload-collection.yaml
