+++ profiles/manifests/hostname-as-podname.yaml
