+++ profiles/manifests/java-ebpf-instrumentations.yaml
