+++ profiles/manifests/query-operation-detector.yaml
