+++ profiles/manifests/semconv.yaml
