+++ profiles/manifests/small-batches.yaml
